Reduce wasted renders
Moved the calculation of the direction to the constructor of the Rating
component.

Having a `setState` inside the `componentDidMount` is not a good thing
to do to begin with, because [it forces a
re-render](https://facebook.github.io/react/docs/react-component.html#componentdidmount).

Since it is not much likely that the text direction of the page will
change during use, I moved the calculation to the constructor.